Make download_size a BigInteger

+++ alembic/versions/2021_06_09_19_00_00-c0e44e063159.py
@@ -22,7 +22,7 @@ def upgrade() -> None:
     op.drop_constraint('downloadevent_version_id_fkey', 'downloadevent', type_='foreignkey')
     op.create_foreign_key('downloadevent_mod_id_fkey', 'downloadevent', 'mod', ['mod_id'], ['id'], ondelete='CASCADE')
     op.create_foreign_key('downloadevent_version_id_fkey', 'downloadevent', 'modversion', ['version_id'], ['id'], ondelete='CASCADE')
-    op.add_column('modversion', sa.Column('download_size', sa.Integer()))
+    op.add_column('modversion', sa.Column('download_size', sa.BigInteger()))
 
 
 def downgrade() -> None:
